PythonProject/sweets-app/backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-jose
passlib[argon2]
argon2-cffi
python-multipart
pydantic-settings